requirements.txt
Django>=3.2
web3
requests
djangorestframework
